changed file names, small corrections
strategies implemented in movement class, fixed motor declarations in moveemnt class, changed file names, added comments to improve understandability

+++ motor.h
@@ -6,7 +6,7 @@ public:
     void forward();
     void backward();
     void stop();
-    void changePwm(int data);
+    void changePwm(float data);
 
 private:
     int motorBackwards;
